splitting application structure
