+++ src/index.js
@@ -3,7 +3,5 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 
-import conf from './json/wedding.json';
-
-ReactDOM.render(<App conf={conf} />, document.getElementById('root'));
+ReactDOM.render(<App />, document.getElementById('root'));
